change the perception about how to update the central node of the DB
This add two interfaces and one class to the logic of the system.

I had a problem when I wanted to keep track of the records in the table
'ClienteProductoAlmacenEmpresa' up to date in real time;
 but then I realized that it would be much easier if, instead,
I implement a message broker/or something that acts as a gateway/queue/filter.
That way I can just make a call to the controller 'PedidoCliente' and, from
there, separate processes.
I mean, that would create a record; and after that, it would take place the
search for the products and whatever logic that manages the stocks.